+++ dwave/cloud/utils/__init__.py
@@ -16,3 +16,7 @@
 
 # imports for backward-compat
 from dwave.cloud.utils.logging import set_loglevel
+
+from dwave.cloud.utils.qubo import (
+    uniform_get, reformat_qubo_as_ising, active_qubits,
+)
